refactor get to use subcommands instead of a manual case statement
